Feature: スクリプトディレクトリへの移行とコード生成機能の追加
- go.workファイルを更新し、scriptsディレクトリを使用
- 新しいスクリプトコマンドを追加し、コード生成機能を実装
- ファイル操作、プロンプト、文字列操作のためのパッケージを作成
- サービスとコンポーネントのテンプレート生成機能を追加
- 不要なtoolディレクトリとutilモジュールを削除
