+++ go.work
@@ -1,6 +1,5 @@
 go 1.22.0
 
 use (
-    ./scripts/tool
-    ./scripts/util
+    ./scripts
 )
