--- scripts/tool/go.mod
@@ -1,23 +1,0 @@
-module tool
-
-go 1.22.0
-
-require (
-	github.com/AlecAivazis/survey/v2 v2.3.7
-	github.com/spf13/cobra v1.10.1
-	util v0.0.0
-)
-
-require (
-	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/kballard/go-shellquote v0.0.0-20180428030007-95032a82bc51 // indirect
-	github.com/mattn/go-colorable v0.1.2 // indirect
-	github.com/mattn/go-isatty v0.0.8 // indirect
-	github.com/mgutz/ansi v0.0.0-20170206155736-9520e82c474b // indirect
-	github.com/spf13/pflag v1.0.9 // indirect
-	golang.org/x/sys v0.0.0-20220722155257-8c9f86f7a55f // indirect
-	golang.org/x/term v0.0.0-20210927222741-03fcf44c2211 // indirect
-	golang.org/x/text v0.4.0 // indirect
-)
-
-replace util => ../util
