--- scripts/util/go.mod
@@ -1,3 +1,0 @@
-module util
-
-go 1.22.0
